# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: 家庭 AI 学习助手 API
  version: 0.6.0
  description: Self-hosted Household account/session and learning contract vertical slices.
servers:
  - url: http://localhost:8000
paths:
...
  /households/{household_id}/children:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
    get:
      operationId: listChildren
      summary: List children in the caller's household
      security:
        - SessionCookie: []
        - BearerSession: []
      responses:
        '200':
          description: Children visible to the caller.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ChildProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: createChild
      summary: Create a child profile in the caller's household
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateChildRequest'
      responses:
        '201':
          description: Child profile created.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildProfile'
        '200':
          description: Idempotent replay of an earlier create request.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/InvalidRequest'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
  /households/{household_id}/children/{child_id}:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
      - name: child_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getChild
      summary: Get a child profile in the caller's household
      security:
        - SessionCookie: []
        - BearerSession: []
      responses:
        '200':
          description: Child profile.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChildProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteChild
      summary: Delete a child profile after cascading Capture object deletion
      description: Local/CI synthetic deletion only. The profile is not removed when any Capture object deletion fails; the same idempotency key may be retried.
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '204':
          description: Child profile and all claimed Capture objects were deleted.
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/InvalidRequest'
        '503':
          description: Deletion is incomplete or temporarily unavailable; retry is safe.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /households/{household_id}/devices:
    parameters:
      - $ref: '#/components/parameters/HouseholdId'
    get:
      operationId: listDevices
      summary: List devices in the caller's household
      security:
        - SessionCookie: []
        - BearerSession: []
      responses:
        '200':
          description: Devices visible to the caller.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      operationId: registerDevice
      summary: Register a device in the caller's household
      security:
        - SessionCookie: []
        - BearerSession: []
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateDeviceRequest'
      responses:
        '201':
          description: Device registered.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '200':
          description: Idempotent replay of an earlier registration.
          headers:
            Idempotency-Replayed:
              schema:
                type: string
                const: 'true'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Device'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/InvalidRequest'
        '409':
          $ref: '#/components/responses/IdempotencyConflict'
...
components:
  parameters:
    HouseholdId:
      name: household_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 8
        maxLength: 128
...
  securitySchemes:
    BearerSession:
      type: http
      scheme: bearer
      bearerFormat: opaque 256-bit session token
      description: Flutter session token; only a SHA-256 digest is stored by the API.
    SessionCookie:
      type: apiKey
      in: cookie
      name: study_session
      description: HttpOnly SameSite=Lax Web session cookie.
  schemas:
...
    CreateChildRequest:
      type: object
      additionalProperties: false
      required: [display_name, grade, curriculum_version, subjects]
      properties:
        display_name:
          type: string
          minLength: 1
          maxLength: 80
        grade:
          type: integer
          minimum: 1
          maximum: 6
        curriculum_version:
          type: string
          minLength: 1
          maxLength: 80
        subjects:
          type: array
          minItems: 1
          items:
            type: string
            enum: [math]
    ChildProfile:
      type: object
      additionalProperties: false
      required: [id, household_id, display_name, grade, curriculum_version, subjects, created_at]
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        display_name:
          type: string
        grade:
          type: integer
        curriculum_version:
          type: string
        subjects:
          type: array
          items:
            type: string
            enum: [math]
        created_at:
          type: string
          format: date-time
    CreateDeviceRequest:
      type: object
      additionalProperties: false
      required: [kind, platform, display_name]
      properties:
        kind:
          type: string
          enum: [child, parent]
        platform:
          type: string
          enum: [ios, android, web]
        display_name:
          type: string
          minLength: 1
          maxLength: 80
    Device:
      type: object
      additionalProperties: false
      required: [id, household_id, kind, platform, display_name, status, registered_at]
      properties:
        id:
          type: string
          format: uuid
        household_id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [child, parent]
        platform:
          type: string
          enum: [ios, android, web]
        display_name:
          type: string
        status:
          type: string
          const: active
        registered_at:
          type: string
          format: date-time
...
  responses:
    Unauthorized:
      description: Missing or invalid synthetic principal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Principal role cannot perform this operation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Resource is not visible in the caller's household.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    IdempotencyConflict:
      description: Idempotency key was reused with a different payload.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InvalidRequest:
      description: Request failed schema validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...